+++ inst/tinytest/testCheck.R
@@ -15,6 +15,20 @@ expect_error(
   addFeatures(study, features = NULL)
 )
 
+# Confirm warning if non-character columns passed in features table
+nonCharacterFeatures <- list(
+  default = data.frame(
+    x = 1:3,
+    y = letters[1:3],
+    z = as.factor(1:3),
+    stringsAsFactors = FALSE
+  )
+)
+expect_warning_xl(
+  addFeatures(study, features = nonCharacterFeatures),
+  ".+non-character.+x.+z"
+)
+
 expect_error(
   addSamples(study, samples = NULL)
 )